+++ src/components/genre-screen/genre-screen.jsx
