+++ src/components/lose-screen/lose-screen.jsx
